requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-simple-history==3.7.0
Pillow==10.4.0
drf-yasg==1.21.7

